requirements.txt
torch==1.12.0+cu113
torchvision==0.13.0+cu113
transformers==4.10.0
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.5
Pillow==8.4.0
matplotlib==3.5.1
